When resolving types, we only need a NumPy array of the argument types. We do not need to fully materialize the array to perform the type checking.
The issue we encountered when we fully materialized the array is as follows:

We need to resolve the types when a Thunk operation is performed (i.e. a * a is called). If we then materialize the sub-arrays of the thunk, we allocate a huge amount of memory when there are many thunks. This memory will not be used for a long time.

Consider, for example, the following chain of operations.

b = a * a * a * a * a

In this operation we create 4 thunks. If we allocate space for the result of each of these thunks we allocate a huge amount of memory that is not immediately necessary, resulting in very bad performance if the arrays are big.

Instead, we simply have 'ready made' NumPy arrays for each possible type. When we need to do type resolution, we simply use one of these ready made arrays. This way we also do not need to do any allocation/deallocation of temporary arrays.